chore: changes after code review
